requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
aiosqlite==0.17.0
    # via tortoise-orm
asgiref==3.4.1
    # via uvicorn
black==21.11b1
    # via -r requirements.in
click==8.0.3
    # via
    #   black
    #   uvicorn
coverage==6.1.2
    # via -r requirements.in
dnspython==2.1.0
    # via email-validator
email-validator==1.1.3
    # via wtforms
flake8==4.0.1
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
idna==3.3
    # via email-validator
iso8601==0.1.16
    # via tortoise-orm
isort==5.10.1
    # via -r requirements.in
markupsafe==2.0.1
    # via wtforms
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
platformdirs==2.4.0
    # via black
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pypika-tortoise==0.1.1
    # via tortoise-orm
pytz==2021.3
    # via tortoise-orm
regex==2021.11.10
    # via black
tomli==1.2.2
    # via black
tortoise-orm==0.17.8
    # via -r requirements.in
typing-extensions==4.0.0
    # via
    #   aiosqlite
    #   black
uvicorn==0.15.0
    # via -r requirements.in
wtforms[email]==3.0.0
    # via -r requirements.in
